Enforce type safety on channel access.
This commit adds in type information into channel access tokens and views. Type information is used to ensure get and set access on a channel via a token always follows the defined type of the token. This benefits the users of the channel data store in that mis-matched type access is no longer possible within component logic.